ufc_project/requirements.txt
beautifulsoup4==4.10.0
ipywidgets==7.6.5
numpy==1.22.0
pandas==1.0.5
PyYAML==6.0
requests==2.31.0
tqdm==4.56.0
mistune>=2.0.3 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
ipython>=8.10.0 # not directly required, pinned by Snyk to avoid a vulnerability
tornado>=6.3.2 # not directly required, pinned by Snyk to avoid a vulnerability
jupyter-server>=2.11.2 # not directly required, pinned by Snyk to avoid a vulnerability
tornado>=6.3.3 # not directly required, pinned by Snyk to avoid a vulnerability
jupyter-server>=2.7.2 # not directly required, pinned by Snyk to avoid a vulnerability
scikit-learn
xgboost
selenium
httpx>=0.24,<0.26
supabase==2.3.1